requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
aiohttp==3.10.11
beautifulsoup4==4.12.3
lxml==5.3.0
